Merge Cosmos master into Aura Cosmos repo.

+++ source/Cosmos.System2/FileSystem/FAT/FatFileSystem.cs
@@ -496,9 +496,7 @@ namespace Cosmos.System.FileSystem.FAT
             byte[] xData;
 
             Read(aCluster, out xData);
-            //TODO: .Net Core
-            //Array.Copy(aData, 0, xData, aOffset, aData.Length);
-            Array.Copy(aData, 0, xData, (int)aOffset, aData.Length);
+            Array.Copy(aData, 0, xData, aOffset, aData.Length);
 
             if (mFatType == FatTypeEnum.Fat32)
             {
